Allow the user to listen to an individual song
This makes the project somewhat "feature complete" because I can do
all of the things that I want to do with it. Of course there are other
features that would be nice but I feel like I could use the project at
this point.

Once concern that I do have is that since I switched from using a list
to store the songs to using a map, the random song is now O(n) instead
of O(1), which might be a problem when I use this in "production" with
almost a thousand songs. I might go back and change that if I can
notice a pause when using it.